adjusting components moved to respective ModOption classes
