+++ SMLHelper/Patchers/OptionsPanelPatcher.cs
@@ -2,7 +2,6 @@
 {
     using Harmony;
     using Options;
-    using System;
     using System.IO;
     using System.Reflection;
     using System.Collections;
@@ -14,14 +13,11 @@
     using UnityEngine.UI;
     using QModManager.API;
 
-    using Object = UnityEngine.Object;
-
     internal class OptionsPanelPatcher
     {
         internal static SortedList<string, ModOptions> modOptions = new SortedList<string, ModOptions>();
 
         private static int  modsTabIndex = -1;
-        private static bool isMainMenu = true; // is options was opened in main menu or in game
 
         internal static void Patch(HarmonyInstance harmony)
         {
@@ -30,10 +26,7 @@
             if (QModServices.Main.FindModById("ModsOptionsAdjusted")?.Enable ?? false)
                 V2.Logger.Log("ModOptionsAdjuster is not inited (ModsOptionsAdjusted mod is active)", LogLevel.Warn);
             else
-            {
-                PatchUtils.PatchClass(harmony, typeof(ModOptionsAdjuster));
                 PatchUtils.PatchClass(harmony, typeof(ModOptionsHeadingsToggle));
-            }
         }
 
 
@@ -51,7 +44,6 @@
         internal static void AddTabs_Postfix(uGUI_OptionsPanel __instance)
         {
             uGUI_OptionsPanel optionsPanel = __instance;
-            isMainMenu = (optionsPanel.GetComponent<MainMenuOptions>() != null);
 
             // Start the modsTab index at a value of -1
             int modsTab = -1;
@@ -86,166 +78,6 @@
             // adding all other options here
             modOptions.Values.ForEach(options => options.AddOptionsToPanel(optionsPanel, modsTab));
         }
-
-        // Adjusting mod options ui elements so they don't overlap with their text labels.
-        // We add corresponding 'adjuster' components to each ui element in mod options tab.
-        // Reason for using components is to skip one frame before manually adjust ui elements to make sure that Unity UI Layout components is updated
-        private static class ModOptionsAdjuster
-        {
-            private static readonly Tuple<string, Type>[] optionTypes = new Tuple<string, Type>[]
-            {
-                Tuple.Create("uGUI_ToggleOption",  typeof(AdjustToggleOption)),
-                Tuple.Create("uGUI_SliderOption",  typeof(AdjustSliderOption)),
-                Tuple.Create("uGUI_ChoiceOption",  typeof(AdjustChoiceOption)),
-                Tuple.Create("uGUI_BindingOption", typeof(AdjustBindingOption))
-            };
-
-            [HarmonyPostfix]
-            [HarmonyPatch(typeof(uGUI_TabbedControlsPanel), nameof(uGUI_TabbedControlsPanel.AddItem), new Type[] { typeof(int), typeof(GameObject) })]
-            private static void AddItem_Postfix(int tabIndex, GameObject __result)
-            {
-                if (__result == null || tabIndex != modsTabIndex)
-                    return;
-
-                foreach (var type in optionTypes)
-                {
-                    if (__result.name.Contains(type.Item1))
-                    {
-                        __result.EnsureComponent(type.Item2);
-                        break;
-                    }
-                }
-            }
-
-            // base class for 'adjuster' components
-            // we add ContentSizeFitter component to text label so it will change width in its Update() based on text
-            // that's another reason to skip one frame
-            private abstract class AdjustModOption: MonoBehaviour
-            {
-                private const float minCaptionWidth_MainMenu = 480f;
-                private const float minCaptionWidth_InGame   = 360f;
-                private GameObject caption = null;
-
-                protected float CaptionWidth { get => caption?.GetComponent<RectTransform>().rect.width ?? 0f; }
-
-                protected static Vector2 SetVec2x(Vector2 vec, float val)  { vec.x = val; return vec; }
-
-                protected void SetCaptionGameObject(string gameObjectPath)
-                {
-                    caption = gameObject.transform.Find(gameObjectPath)?.gameObject;
-
-                    if (!caption)
-                    {
-                        V2.Logger.Log($"AdjustModOption: caption gameobject '{gameObjectPath}' not found", V2.LogLevel.Warn);
-                        return;
-                    }
-
-                    caption.AddComponent<LayoutElement>().minWidth = isMainMenu? minCaptionWidth_MainMenu: minCaptionWidth_InGame;
-                    caption.AddComponent<ContentSizeFitter>().horizontalFit = ContentSizeFitter.FitMode.PreferredSize; // for autosizing captions
-
-                    RectTransform transform = caption.GetComponent<RectTransform>();
-                    transform.SetAsFirstSibling(); // for HorizontalLayoutGroup
-                    transform.pivot = SetVec2x(transform.pivot, 0f);
-                    transform.anchoredPosition = SetVec2x(transform.anchoredPosition, 0f);
-                }
-            }
-
-            // in case of ToggleOption there is no need to manually move elements
-            // other option types don't work well with HorizontalLayoutGroup :(
-            private class AdjustToggleOption: AdjustModOption
-            {
-                private const float spacing = 20f;
-
-                public void Awake()
-                {
-                    HorizontalLayoutGroup hlg = gameObject.transform.Find("Toggle").gameObject.AddComponent<HorizontalLayoutGroup>();
-                    hlg.childControlWidth = false;
-                    hlg.childForceExpandWidth = false;
-                    hlg.spacing = spacing;
-
-                    SetCaptionGameObject("Toggle/Caption");
-
-                    Destroy(this);
-                }
-            }
-
-            private class AdjustSliderOption: AdjustModOption
-            {
-                private const float spacing = 25f;
-                private const float sliderValueWidth = 85f;
-
-                public IEnumerator Start()
-                {
-                    SetCaptionGameObject("Slider/Caption");
-                    yield return null; // skip one frame
-
-                    // for some reason sliders don't update their handle positions sometimes
-                    uGUI_SnappingSlider slider = gameObject.GetComponentInChildren<uGUI_SnappingSlider>();
-                    AccessTools.Method(typeof(Slider), "UpdateVisuals")?.Invoke(slider, null);
-
-                    // changing width for slider value label
-                    RectTransform sliderValueRect = gameObject.transform.Find("Slider/Value") as RectTransform;
-                    sliderValueRect.sizeDelta = SetVec2x(sliderValueRect.sizeDelta, sliderValueWidth);
-
-                    // changing width for slider
-                    RectTransform rect = gameObject.transform.Find("Slider/Background") as RectTransform;
-
-                    float widthAll = gameObject.GetComponent<RectTransform>().rect.width;
-                    float widthSlider = rect.rect.width;
-                    float widthText = CaptionWidth + spacing;
-
-                    if (widthText + widthSlider + sliderValueWidth > widthAll)
-                        rect.sizeDelta = SetVec2x(rect.sizeDelta, widthAll - widthText - sliderValueWidth - widthSlider);
-
-                    Destroy(this);
-                }
-            }
-
-            private class AdjustChoiceOption: AdjustModOption
-            {
-                private const float spacing = 10f;
-
-                public IEnumerator Start()
-                {
-                    SetCaptionGameObject("Choice/Caption");
-                    yield return null; // skip one frame
-
-                    RectTransform rect = gameObject.transform.Find("Choice/Background") as RectTransform;
-
-                    float widthAll = gameObject.GetComponent<RectTransform>().rect.width;
-                    float widthChoice = rect.rect.width;
-                    float widthText = CaptionWidth + spacing;
-
-                    if (widthText + widthChoice > widthAll)
-                        rect.sizeDelta = SetVec2x(rect.sizeDelta, widthAll - widthText - widthChoice);
-
-                    Destroy(this);
-                }
-            }
-
-            private class AdjustBindingOption: AdjustModOption
-            {
-                private const float spacing = 10f;
-
-                public IEnumerator Start()
-                {
-                    SetCaptionGameObject("Caption");
-                    yield return null; // skip one frame
-
-                    RectTransform rect = gameObject.transform.Find("Bindings") as RectTransform;
-
-                    float widthAll = gameObject.GetComponent<RectTransform>().rect.width;
-                    float widthBinding = rect.rect.width;
-                    float widthText = CaptionWidth + spacing;
-
-                    if (widthText + widthBinding > widthAll)
-                        rect.sizeDelta = SetVec2x(rect.sizeDelta, widthAll - widthText - widthBinding);
-
-                    Destroy(this);
-                }
-            }
-        }
-
 
         // Class for collapsing/expanding options in 'Mods' tab
         // Options can be collapsed/expanded by clicking on mod's title or arrow button
